pyproject.toml
[tool.poetry]
name = "up42-py"
version = "0.37.0a12"
description = "Python SDK for UP42, the geospatial marketplace and developer platform."
authors = ["UP42 GmbH <support@up42.com>"]
license = "https://github.com/up42/up42-py/blob/master/LICENSE"
readme = "README.md"
packages = [
    { include = "up42", from = "." },
]
classifiers=[
        "Programming Language :: Python :: 3",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
        "Development Status :: 4 - Beta",
        "Intended Audience :: Developers",
    ]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
requests =  "^2.31.0"
requests-oauthlib = "^1.3.1"
tenacity = "^8.2.2"
tqdm = "^4.66.0"
geojson = "^3.0.1"
geopandas = "^0.13.2"
pystac-client = "^0.7.2"


[tool.poetry.dev-dependencies]
nbconvert = "^6.5.4"
black = "^22.10.0"
requests-mock = "^1.9.3"
pylint = "^3.0.3"
pytest = "^7.2.0"
pytest-pylint = "0.21.0"
pytest-sugar = "^0.9.4"
mypy = "^1.8.0"
python-dateutil = "^2.8.2"
types-python-dateutil = "^2.8.19.4"
types-requests = "^2.28.11.5"
types-mock = "^4.0.15.2"
types-click = "^7.1.8"
mypy-extensions = "^1.0.0"
pytest-cov = "^4.0.0"
pytest-mypy = "^0.10.2"
mock = "^4.0.3"
pre-commit = "^3.5.0"
types-tqdm = "^4.66.0.20240106"
pandas-stubs = "^2.0.1.230501"
types-oauthlib = "^3.2.0.20240217"
pylint-google-style-guide-imports-enforcing = "^1.3.0"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^8.5.11"
mkdocstrings = ">=0.19.0"
mkdocstrings-python = ">=1.3.0"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.22.0"
mkdocs-autolinks-plugin = ">=0.6.0"
mkdocs-macros-plugin = ">=1.0.0"


[tool.poetry.group.viz.dependencies]
rasterio = "^1.3.8"
folium = "^0.14.0"
branca = "^0.6.0"
matplotlib = "^3.7.2"
descartes = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
